+++ install-standalone.sh
@@ -17,9 +17,10 @@ git clone https://github.com/xtr4nge/FruityC2
 # SSL
 echo
 echo "--------------------------------"
-echo "Create SSL certificate (default)"
+echo "Create SSL certificate (FruityC2)"
 echo "--------------------------------"
-openssl req -x509 -nodes -days 365 -newkey rsa:2048 -keyout FruityC2/certs/nginx.key -out FruityC2/certs/nginx.crt
+openssl req -x509 -nodes -days 365 -newkey rsa:2048 -keyout certs/fruityc2.key -out certs/fruityc2.crt
+cat certs/fruityc2.key certs/fruityc2.crt > certs/fruityc2.pem
 
 echo
 echo "--------------------------------"