+++ install.sh
@@ -9,9 +9,10 @@ pip install Flask-CORS
 
 # SSL
 echo "--------------------------------"
-echo "Create SSL certificate (default)"
+echo "Create SSL certificate (FruityC2)"
 echo "--------------------------------"
-openssl req -x509 -nodes -days 365 -newkey rsa:2048 -keyout certs/nginx.key -out certs/nginx.crt
+openssl req -x509 -nodes -days 365 -newkey rsa:2048 -keyout certs/fruityc2.key -out certs/fruityc2.crt
+cat certs/fruityc2.key certs/fruityc2.crt > certs/fruityc2.pem
 
 echo
 echo "FruityC2: https://github.com/xtr4nge/FruityC2"