+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 ==============
 
+[v0.6](https://github.com/xtr4nge/FruityC2/releases/tag/v0.6)
+----------------
+- HTTP Listener issue has been fixed
+- FruityC2 Client SSL has been implemented
+- Default|Custom SSL certificate (FruityC2) has been implemented
+- Default|Custom SSL certificate (Listener) has been implemented
+
 
 [v0.5](https://github.com/xtr4nge/FruityC2/releases/tag/v0.5)
 ----------------